+++ lib/composer/composer/autoload_classmap.php
@@ -561,6 +561,8 @@ return array(
     'OCP\\Mail\\IEMailTemplate' => $baseDir . '/lib/public/Mail/IEMailTemplate.php',
     'OCP\\Mail\\IMailer' => $baseDir . '/lib/public/Mail/IMailer.php',
     'OCP\\Mail\\IMessage' => $baseDir . '/lib/public/Mail/IMessage.php',
+    'OCP\\Mail\\Provider\\Address' => $baseDir . '/lib/public/Mail/Provider/Address.php',
+    'OCP\\Mail\\Provider\\Attachment' => $baseDir . '/lib/public/Mail/Provider/Attachment.php',
     'OCP\\Mail\\Provider\\IAddress' => $baseDir . '/lib/public/Mail/Provider/IAddress.php',
     'OCP\\Mail\\Provider\\IAttachment' => $baseDir . '/lib/public/Mail/Provider/IAttachment.php',
     'OCP\\Mail\\Provider\\IManager' => $baseDir . '/lib/public/Mail/Provider/IManager.php',
@@ -570,6 +572,7 @@ return array(
     'OCP\\Mail\\Provider\\IService' => $baseDir . '/lib/public/Mail/Provider/IService.php',
     'OCP\\Mail\\Provider\\IServiceIdentity' => $baseDir . '/lib/public/Mail/Provider/IServiceIdentity.php',
     'OCP\\Mail\\Provider\\IServiceLocation' => $baseDir . '/lib/public/Mail/Provider/IServiceLocation.php',
+    'OCP\\Mail\\Provider\\Message' => $baseDir . '/lib/public/Mail/Provider/Message.php',
     'OCP\\Migration\\BigIntMigration' => $baseDir . '/lib/public/Migration/BigIntMigration.php',
     'OCP\\Migration\\IMigrationStep' => $baseDir . '/lib/public/Migration/IMigrationStep.php',
     'OCP\\Migration\\IOutput' => $baseDir . '/lib/public/Migration/IOutput.php',
@@ -1598,10 +1601,7 @@ return array(
     'OC\\Mail\\EMailTemplate' => $baseDir . '/lib/private/Mail/EMailTemplate.php',
     'OC\\Mail\\Mailer' => $baseDir . '/lib/private/Mail/Mailer.php',
     'OC\\Mail\\Message' => $baseDir . '/lib/private/Mail/Message.php',
-    'OC\\Mail\\Provider\\Address' => $baseDir . '/lib/private/Mail/Provider/Address.php',
-    'OC\\Mail\\Provider\\Attachment' => $baseDir . '/lib/private/Mail/Provider/Attachment.php',
     'OC\\Mail\\Provider\\Manager' => $baseDir . '/lib/private/Mail/Provider/Manager.php',
-    'OC\\Mail\\Provider\\Message' => $baseDir . '/lib/private/Mail/Provider/Message.php',
     'OC\\Memcache\\APCu' => $baseDir . '/lib/private/Memcache/APCu.php',
     'OC\\Memcache\\ArrayCache' => $baseDir . '/lib/private/Memcache/ArrayCache.php',
     'OC\\Memcache\\CADTrait' => $baseDir . '/lib/private/Memcache/CADTrait.php',
