+++ lib/public/Mail/Provider/IAddress.php
@@ -25,18 +25,46 @@ declare(strict_types=1);
 namespace OCP\Mail\Provider;
 
 /**
- * Interface IMessage
- *
+ * Mail Address Interface
+ * 
+ * This interface is a base requirement of methods and functionality used to construct a mail address object
+ * 
  * @since 30.0.0
  */
 interface IAddress {
 	
+	/**
+	 * sets the mail address
+	 * 
+	 * @since 30.0.0
+	 * @param string $value     mail address (test@example.com)
+	 * @return self				returns the current object
+	 */
 	public function setAddress(string $value): self;
 
-	public function getAddress(): string;
+	/**
+	 * gets the mail address
+	 * 
+	 * @since 30.0.0
+	 * @return string			returns the mail address
+	 */
+	public function getAddress(): string | null;
 
-	public function setName(string $value): self;
+	/**
+	 * sets the mail address label/name
+	 * 
+	 * @since 30.0.0
+	 * @param string $value     mail address label/name
+	 * @return self				returns the current object
+	 */
+	public function setLabel(string $value): self;
 
-	public function getName(): string;
+	/**
+	 * gets the mail address label/name
+	 * 
+	 * @since 30.0.0
+	 * @return string			returns the mail address label/name
+	 */
+	public function getLabel(): string | null;
 
 }
