+++ lib/public/Mail/Provider/IProvider.php
@@ -27,7 +27,7 @@ namespace OCP\Mail\Provider;
 /**
  * Mail Provider Interface
  * 
- * This interface is used for defininig a mail provider
+ * This interface is a base requirement of methods and functionality used to construct a mail provider object
  * 
  * @since 30.0.0
  */
@@ -37,7 +37,7 @@ interface IProvider {
 	 * An arbitrary unique text string identifying this provider
 	 * 
 	 * @since 30.0.0
-	 * @return string			id of this provider (e.g. UUID or 'IMAP/SMTP' or anything else)
+	 * @return string				id of this provider (e.g. UUID or 'IMAP/SMTP' or anything else)
 	 */
 	public function id(): string;
 
@@ -45,7 +45,7 @@ interface IProvider {
 	 * The localized human frendly name of this provider
 	 * 
 	 * @since 30.0.0
-	 * @return string			label/name of this provider (e.g. Plain Old IMAP/SMTP)
+	 * @return string				label/name of this provider (e.g. Plain Old IMAP/SMTP)
 	 */
 	public function label(): string;
 
@@ -53,7 +53,7 @@ interface IProvider {
 	 * Determain if any services are configured for a specific user
 	 * 
 	 * @since 30.0.0
-	 * @return bool 			true if any services are configure for the user
+	 * @return bool 				true if any services are configure for the user
 	 */
 	public function hasServices(string $uid): bool;
 
@@ -61,7 +61,7 @@ interface IProvider {
 	 * retrieve collection of services for a specific user
 	 * 
 	 * @since 30.0.0
-	 * @return array<IService>	collection of service objects
+	 * @return array<int, IService>		collection of service objects
 	 */
 	public function listServices(string $uid): array;
 
@@ -71,6 +71,7 @@ interface IProvider {
 	 * @since 30.0.0
 	 * @param string $uid			user id of user to configure service for
 	 * @param IService $service 	service configuration object
+	 * @return string				id of created service
 	 */
 	public function createService(string $uid, IService $service): string;
 
@@ -80,6 +81,7 @@ interface IProvider {
 	 * @since 30.0.0
 	 * @param string $uid			user id of user to configure service for
 	 * @param IService $service 	service configuration object
+	 * @return string				id of modifided service
 	 */
 	public function modifyService(string $uid, IService $service): string;
 
@@ -89,6 +91,7 @@ interface IProvider {
 	 * @since 30.0.0
 	 * @param string $uid			user id of user to delete service for
 	 * @param IService $service 	service configuration object
+	 * @return bool					status of delete action
 	 */
 	public function deleteService(string $uid, IService $service): bool;
