+++ lib/public/Mail/Provider/IServiceIdentity.php
@@ -27,16 +27,20 @@ namespace OCP\Mail\Provider;
 use RuntimeException;
 
 /**
- * This is the interface that is implemented by apps that
- * implement a mail provider
- * @since 30
+ * Mail Service Identity Interface
+ * 
+ * This interface is a base requirement of methods and functionality used to construct a mail service identity.
+ * 
+ * @since 30.0.0
  */
 interface IServiceIdentity {
 
 	/**
 	 * An arbitrary unique text string identifying this credential type
+	 * 
 	 * @since 30.0.0
+	 * @return string
 	 */
-	public function id(): string;
+	public function type(): string;
 
 }
