+++ lib/public/Mail/Provider/IServiceLocation.php
@@ -27,15 +27,19 @@ namespace OCP\Mail\Provider;
 use RuntimeException;
 
 /**
- * This is the interface that is implemented by apps that
- * implement a mail provider
+ * Mail Service Location Interface
+ * 
+ * This interface is a base requirement of methods and functionality used to construct a mail service location.
+ * 
  * @since 30.0.0
  */
 interface IServiceLocation {
 
 	/**
+	 * A string identifiing this location type
 	 * 
 	 * @since 30.0.0
+	 * @return string
 	 */
 	public function type(): string;
